pyproject.toml
[tool.poetry]
name = "nonebot-plugin-bilirequestall"
version = "0.3.0"
description = "以B站UID审核入群 😎"
authors = ["Shadow403 <admin@shadow403.cn>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0"
nonebot2 = ">=2.0.0b1"
httpx = ">=0.27.0"
pydantic = ">=2.8.0"
nonebot-adapter-onebot = ">=2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
